Rename to fit module naming pattern
